Fix more (MSVC) compiler warnings
reg. `size_t` vs. `int`, still not all (and some of these changes result in new warnings elsewhere), but 496 less than before (still 452 left).

+++ src/engraving/libmscore/stringdata.h
@@ -72,7 +72,7 @@ public:
     void        fretChords(Chord* chord) const;
     int         getPitch(int string, int fret, Staff* staff) const;
     static int  pitchOffsetAt(Staff* staff);
-    int         strings() const { return stringTable.size(); }
+    size_t      strings() const { return stringTable.size(); }
     int         frettedStrings() const;
     const std::vector<instrString>& stringList() const { return stringTable; }
     std::vector<instrString>& stringList() { return stringTable; }
